Moved models to shared folder
