+++ src/app/components/cartracker/link-to-user/link-to-user.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Vehicle } from 'src/app/models/vehicle.model';
+import { Vehicle } from 'src/app/shared/models/vehicle.model';
 import { ClientUser } from 'src/app/shared/models/clientUser';
 import { UserService } from 'src/app/shared/services/user.service';
 import { VehicleService } from 'src/app/shared/services/vehicle.service';
