+++ src/app/shared/models/locationPoint.model.ts
