+++ src/app/shared/models/nawinformation.model.ts
