+++ src/app/shared/models/vehicle.model.ts
